requirements.txt
asgiref==3.5.0
astroid==2.11.3
black==22.3.0
cfgv==3.3.1
click==8.1.2
dill==0.3.4
distlib==0.3.4
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
gunicorn==20.1.0
identify==2.4.12
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy==0.942
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.18.1
PyJWT==2.3.0
pylint==2.13.6
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
virtualenv==20.14.1
wrapt==1.14.0
django-cors-headers==3.11.0
psycopg2-binary==2.8.6